Release v1.2.10

Core
 - Release v1.2.10
 - github: update build.yml
 - aseqdump: Add UMP support
 - nhlt: add nhlt-dmic-info utility
ALSA Control (alsactl)
 - alsactl: fix compilation when building in a subdir
 - alsactl: fix the verbose compilation warnings for latest gcc
 - alsactl: fix the copy-n-paste typo (SND_RAWMIDI_STREAM_*)
 - alsactl: add define to compile with glibc 2.38
ALSA RawMidi Utility (amidi)
 - reshuffle included files to include config.h as first
 - amidi: fix the verbose compilation warnings for latest gcc
Audio Transfer utility
 - reshuffle included files to include config.h as first
 - axfer: fix the verbose compilation warnings for latest gcc
NHLT ACPI parser
 - nhlt-dmic-info: fix the verbose compilation warnings for latest gcc
 - nhlt: add nhlt-dmic-info utility
Speaker Test
 - speaker-test: fix the verbose compilation warnings for latest gcc
 - speaker-test: allow large buffer and period time setup - up to 100 seconds
aconnect
 - aconnect: fix the verbose compilation warnings for latest gcc
 - aconnect: Add UMP support
alsaloop
 - reshuffle included files to include config.h as first
 - alsaloop: fix the verbose compilation warnings for latest gcc
alsamixer
 - alsamixer: fix the verbose compilation warnings for latest gcc
alsatplg (topology)
 - topology: fix the verbose compilation warnings for latest gcc
 - topology: plugins: nhlt: remove dmic error print
 - Topology: NHLT: Intel: Update DMIC FIR coefficients
 - topology: nhlt: intel: support more device types and directions
 - topology: pre-processor: Add support for CombineArrays
 - topology: plugins: nhlt: set dmic stereo mode only in hw version 1
alsaucm
 - alsaucm: fix the verbose compilation warnings for latest gcc
amixer
 - reshuffle included files to include config.h as first
 - amixer: fix the verbose compilation warnings for latest gcc
aplay/arecord
 - aplay: fix the verbose compilation warnings for latest gcc
aplaymidi/arecordmidi
 - aplaymidi: fix the verbose compilation warnings for latest gcc
 - aplaymidi: Add UMP support
aseqdump
 - aseqdump: fix MIDI 2.0 code - it compiles now
 - aseqdump: fix the verbose compilation warnings for latest gcc
 - aseqdump: Add options to switch view mode
 - aseqdump: Correct wrong channel number
 - aseqdump: Align outputs of UMP MIDI 1.0 with legacy MIDI 1.0
 - aseqdump: Add UMP support
aseqnet
 - aseqnet: fix the verbose compilation warnings for latest gcc
bat (basic audio tester)
 - bat: fix the verbose compilation warnings for latest gcc
gitcompile
 - gitcompile: set more verbose compilation warnings
iecset
 - iecset: fix the verbose compilation warnings for latest gcc

+++ bat/latencytest.c
@@ -13,6 +13,8 @@
  *
  */
 
+#include "aconfig.h"
+
 #include <math.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -131,7 +133,7 @@ static void play_and_listen(struct bat *bat, void *buffer, int frames)
 
 		/* Do not listen to more than a second
 		   Maybe too much background noise */
-		if (bat->latency.samples > bat->rate) {
+		if ((unsigned int)bat->latency.samples > bat->rate) {
 			bat->latency.error++;
 
 			if (bat->latency.error > LATENCY_TEST_NUMBER) {
@@ -191,7 +193,7 @@ int handleinput(struct bat *bat, void *buffer, int frames)
 		bat->latency.samples += frames;
 
 		/* 1 second elapsed */
-		if (bat->latency.samples >= bat->rate) {
+		if ((unsigned int)bat->latency.samples >= bat->rate) {
 			calculate_threshold(bat);
 			bat->latency.state = LATENCY_STATE_PLAY_AND_LISTEN;
 			bat->latency.samples = 0;
@@ -208,7 +210,7 @@ int handleinput(struct bat *bat, void *buffer, int frames)
 	case LATENCY_STATE_WAITING:
 		bat->latency.samples += frames;
 
-		if (bat->latency.samples > bat->rate) {
+		if ((unsigned int)bat->latency.samples > bat->rate) {
 			/* 1 second elapsed, start over */
 			bat->latency.samples = 0;
 			bat->latency.state = LATENCY_STATE_MEASURE_FOR_1_SECOND;
